requirements.txt
absl-py==1.2.0
ale-py==0.8.0
attrs==22.1.0
box2d-py==2.3.5
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.0
contourpy==1.0.5
cycler==0.11.0
Cython==0.29.32
decorator==4.4.2
fasteners==0.18
fonttools==4.37.3
glfw==2.5.5
gym==0.26.1
gym-notices==0.0.8
idna==3.4
imageio==2.22.0
imageio-ffmpeg==0.4.7
importlib-resources==5.9.0
iniconfig==1.1.1
kiwisolver==1.4.4
lz4==4.0.2
matplotlib==3.6.0
moviepy==1.0.3
mujoco==2.2.0
mujoco-py==2.1.2.14
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
proglog==0.1.10
py==1.11.0
pycparser==2.21
pygame==2.1.0
PyOpenGL==3.1.6
pyparsing==3.0.9
pytest==7.0.1
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
swig==4.0.2
tomli==2.0.1
torch==1.12.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
